Add trading fee to trading request
