+++ blockchain/beaconpdeproducer.go
@@ -107,51 +107,6 @@ func isRightRatio(
 	return false
 }
 
-func addShareAmountUpV2(
-	beaconHeight uint64,
-	token1IDStr string,
-	token2IDStr string,
-	contributedTokenIDStr string,
-	contributorAddrStr string,
-	amt uint64,
-	currentPDEState *CurrentPDEState,
-) {
-	pdeShareOnTokenPrefix := string(lvdb.BuildPDESharesKeyV2(beaconHeight, token1IDStr, token2IDStr, ""))
-	totalSharesOnToken := uint64(0)
-	for key, value := range currentPDEState.PDEShares {
-		if strings.Contains(key, pdeShareOnTokenPrefix) {
-			totalSharesOnToken += value
-		}
-	}
-	pdeShareKey := string(lvdb.BuildPDESharesKeyV2(beaconHeight, token1IDStr, token2IDStr, contributorAddrStr))
-	if totalSharesOnToken == 0 {
-		currentPDEState.PDEShares[pdeShareKey] = amt
-		return
-	}
-	poolPairKey := string(lvdb.BuildPDEPoolForPairKey(beaconHeight, token1IDStr, token2IDStr))
-	poolPair, found := currentPDEState.PDEPoolPairs[poolPairKey]
-	if !found || poolPair == nil {
-		currentPDEState.PDEShares[pdeShareKey] = amt
-		return
-	}
-	poolValue := poolPair.Token1PoolValue
-	if poolPair.Token2IDStr == contributedTokenIDStr {
-		poolValue = poolPair.Token2PoolValue
-	}
-	if poolValue == 0 {
-		currentPDEState.PDEShares[pdeShareKey] = amt
-	}
-	increasingAmt := big.NewInt(0)
-	increasingAmt.Mul(big.NewInt(int64(totalSharesOnToken)), big.NewInt(int64(amt)))
-	increasingAmt.Div(increasingAmt, big.NewInt(int64(poolValue)))
-	currentShare, found := currentPDEState.PDEShares[pdeShareKey]
-	addedUpAmt := increasingAmt.Uint64()
-	if found {
-		addedUpAmt += currentShare
-	}
-	currentPDEState.PDEShares[pdeShareKey] = addedUpAmt
-}
-
 func (blockchain *BlockChain) buildInstructionsForPDEContribution(
 	contentStr string,
 	shardID byte,
@@ -250,6 +205,7 @@ func (blockchain *BlockChain) buildInstructionsForPDEContribution(
 		return [][]string{matchedInst}, nil
 	}
 
+// the contribution was not right with the current ratio of the pair
 	delete(currentPDEState.WaitingPDEContributions, waitingContribPairKey)
 	refundInst1 := buildRefundContributionInst(
 		meta.PDEContributionPairID,
@@ -319,7 +275,7 @@ func (blockchain *BlockChain) buildInstructionsForPDETrade(
 	}
 	invariant := big.NewInt(0)
 	invariant.Mul(big.NewInt(int64(tokenPoolValueToSell)), big.NewInt(int64(tokenPoolValueToBuy)))
-	fee := pdeTradeReqAction.Meta.SellAmount / PDEDevisionAmountForFee
+	fee := pdeTradeReqAction.Meta.TradingFee
 	newTokenPoolValueToSell := big.NewInt(0)
 	newTokenPoolValueToSell.Add(big.NewInt(int64(tokenPoolValueToSell)), big.NewInt(int64(pdeTradeReqAction.Meta.SellAmount)))
 	newTokenPoolValueToSellAfterFee := big.NewInt(0).Sub(newTokenPoolValueToSell, big.NewInt(int64(fee)))
